started rewriting for dependancy injection using cocona
