+++ GCS.Core/Helpers/FileHelper.cs
