+++ GCS.Core/Helpers/PathValidator.cs
