+++ GCS.Core/Interfaces/IGameManager.cs
@@ -1,0 +1,10 @@
+namespace GCS.Core
+{
+    public interface IGameDataManager
+    {
+        GameData LoadGameData();
+        void SaveGameData(GameData gameData);
+        void AddGameData(string title, string configPath, List<Profile> profiles);
+        void DeleteGameData(string title);
+    }
+}
