+++ GCS.Core/Models/Game.cs
@@ -1,0 +1,9 @@
+namespace GCS.Core
+{
+    public class Game
+    {
+        public string Title { get; set; }
+        public string configPath { get; set; }
+        public List<Profile> Profiles { get; set; }
+    }
+}
