+++ GCS.Core/Models/Profile.cs
@@ -1,0 +1,8 @@
+namespace GCS.Core
+{
+    public class Profile
+    {
+        public string title { get; set; }
+        public string profilePath { get; set; }
+    }
+}
